refactor: move condition/action factory to trigger
