+++ next/api/src/ticket/automation/index.ts
@@ -1,2 +1,2 @@
-export { Condition, condition } from './condition';
-export { Action, action } from './action';
+export { Condition } from './condition';
+export { Action } from './action';
